Task 5 variable aux now is a unsigned long int

+++ 0x14-bit_manipulation/5-flip_bits.c
@@ -7,9 +7,10 @@
  * @m: second number
  * Return: number of bits to set
  */
+
 unsigned int flip_bits(unsigned long int n, unsigned long int m)
 {
-	unsigned int aux; /*keep xor of n and m*/
+	unsigned long int aux; /*keep xor of n and m*/
 	unsigned int counter_of_difference = 0; /*count the number of diference*/
 
 	aux = (n ^ m); /*the result will be compauxre with aux-1*/
